x/credit/types/message_repay_credits_test.go
package types

import (
	"testing"

	"core/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgRepayCredits_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgRepayCredits
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgRepayCredits{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgRepayCredits{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
